requirements.txt
# Core dependencies
playwright>=1.40.0
python-dotenv>=1.0.0

# AI/ML dependencies (optional)
openai>=1.0.0
anthropic>=0.7.0
google-genai>=1.38.0

# HTTP requests
requests>=2.31.0

# Task scheduling
schedule>=1.2.0

# Data handling
pyyaml>=6.0.1

# FastAPI and backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Browser automation
browser-use>=0.1.0
